cash/automation/requirements.txt
playwright==1.47.0
python-dotenv==1.0.1
pyyaml==6.0.2
openpyxl==3.1.5
schedule==1.2.2
requests==2.32.3
pyrogram==2.0.106
